+++ src/main/java/roomescape/member/auth/controller/dto/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.member.controller.dto;
+package roomescape.member.auth.controller.dto;
 
 public record LoginMember(Long id, String email, String name, String role) {
 }
