+++ src/main/java/roomescape/member/auth/resolver/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.resolver;
+package roomescape.member.auth.resolver;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -8,7 +8,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.member.controller.dto.LoginMember;
+import roomescape.member.auth.controller.dto.LoginMember;
 import roomescape.member.domain.Member;
 import roomescape.member.service.MemberService;
