+++ src/main/java/roomescape/member/controller/AdminViewController.java
@@ -1,0 +1,16 @@
+package roomescape.member.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminViewController {
+
+    @GetMapping
+    public String admin() {
+        return "admin/index";
+    }
+
+}
