+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -1,21 +1,12 @@
 package roomescape.member.controller;
 
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.member.controller.dto.LoginRequest;
+import org.springframework.web.bind.annotation.*;
 import roomescape.member.controller.dto.MemberResponse;
 import roomescape.member.controller.dto.SignupRequest;
 import roomescape.member.service.MemberService;
+
+import java.util.List;
 
 @RestController
 public class MemberController {
@@ -26,53 +17,15 @@ public class MemberController {
         this.memberService = memberService;
     }
 
-    @ResponseStatus(HttpStatus.OK)
-    @GetMapping("/login/check")
-    public MemberResponse loginCheck(HttpServletRequest request) {
-        Cookie[] cookies = request.getCookies();
-        String token = extractTokenFromCookie(cookies)
-                .orElseThrow(() -> new IllegalArgumentException("token 쿠키가 없습니다."));
-        return memberService.search(token);
-    }
-
     @GetMapping("/members")
     public List<MemberResponse> getMembers() {
         return memberService.getAll();
     }
 
     @ResponseStatus(HttpStatus.OK)
-    @PostMapping("/login")
-    public void login(@RequestBody LoginRequest request, HttpServletResponse response) {
-        String accessToken = memberService.login(request);
-        Cookie cookie = new Cookie("token", accessToken);
-        cookie.setHttpOnly(true);
-        cookie.setPath("/");
-        response.addCookie(cookie);
-    }
-
-    @PostMapping("/logout")
-    public void logout(HttpServletRequest request, HttpServletResponse response) {
-        Cookie[] cookies = request.getCookies();
-        extractTokenFromCookie(cookies).ifPresent(token -> {
-            Cookie expiredCookie = new Cookie("token", null);
-            expiredCookie.setMaxAge(0);
-            expiredCookie.setHttpOnly(true);
-            expiredCookie.setPath("/");
-            response.addCookie(expiredCookie);
-        });
-    }
-
-    @ResponseStatus(HttpStatus.OK)
     @PostMapping("/members")
     public MemberResponse signup(@RequestBody SignupRequest request) {
         return memberService.add(request);
-    }
-
-    private Optional<String> extractTokenFromCookie(Cookie[] cookies) {
-        return Arrays.stream(cookies)
-                .filter(cookie -> "token".equals(cookie.getName()))
-                .map(Cookie::getValue)
-                .findFirst();
     }
 
 }
