+++ src/main/java/roomescape/member/controller/UserViewController.java
@@ -1,0 +1,14 @@
+package roomescape.member.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserViewController {
+
+    @GetMapping("/signup")
+    public String signup() {
+        return "signup";
+    }
+
+}
