+++ src/main/java/roomescape/reservation/controller/UserReservationViewController.java
@@ -1,24 +1,14 @@
-package roomescape.view;
+package roomescape.reservation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class UserController {
+public class UserReservationViewController {
 
     @GetMapping("/reservation")
     public String userReservation() {
         return "reservation";
-    }
-
-    @GetMapping("/login")
-    public String login() {
-        return "login";
-    }
-
-    @GetMapping("/signup")
-    public String signup() {
-        return "signup";
     }
 
 }
