+++ src/main/java/roomescape/time/controller/AdminTimeViewController.java
@@ -1,0 +1,16 @@
+package roomescape.time.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/admin")
+public class AdminTimeViewController {
+
+    @GetMapping("/time")
+    public String time() {
+        return "admin/time";
+    }
+
+}
